fix for tests
