+++ TODO.md
@@ -1,2 +1,3 @@
+- update for rhel-8+ nftables, RHEL docs recommend nftables for new firewalls <https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/8/html/configuring_and_managing_networking/getting-started-with-nftables_configuring-and-managing-networking>
 - iptables' `-S` not supported in libraries/provider_firewall_iptables.rb
 - save action might not make sense for firewalls
