+++ libraries/provider_firewall_iptables.rb
@@ -23,7 +23,7 @@ class Chef
     include FirewallCookbook::Helpers::Iptables
 
     provides :firewall, os: 'linux', platform_family: %w(rhel fedora amazon) do |node|
-      node['platform_version'].to_f < 7.0 || node['firewall']['redhat7_iptables']
+      (node['platform_version'].to_i < 7 && !amazon_linux?) || node['platform_version'].to_i >= 8 || node['firewall']['redhat7_iptables']
     end
 
     def whyrun_supported?
