+++ test/fixtures/cookbooks/firewall-test/recipes/default.rb
@@ -24,7 +24,7 @@ end
 firewall_rule 'addremove' do
   port 1236
   command :allow
-  only_if { rhel? || node['firewall']['ubuntu_iptables'] } # don't do this on ufw, will reset ufw on every converge
+  only_if { rhel? || amazon_linux? || node['firewall']['ubuntu_iptables'] } # don't do this on ufw, will reset ufw on every converge
 end
 
 firewall_rule 'addremove2' do
@@ -35,7 +35,7 @@ end
 firewall_rule 'protocolnum' do
   protocol 112
   command :allow
-  only_if { rhel? || node['firewall']['ubuntu_iptables'] } # debian ufw doesn't support protocol numbers
+  only_if { rhel? || amazon_linux? || node['firewall']['ubuntu_iptables'] } # debian ufw doesn't support protocol numbers
 end
 
 firewall_rule 'prepend' do
