--- test/integration/default/spec_helper.rb
@@ -1,8 +1,0 @@
-# helpers
-def firewalld?
-  os.redhat? && os[:release].to_f >= 7.0
-end
-
-def iptables?
-  (os.redhat? && os[:release].to_f < 7.0)
-end
