+++ test/integration/windows/inspec/windows_spec.rb
